Reformat purl stored in "data.identity"
The purl stored in data.identity of Eiffel event was of the form:
"pkg:<build_path>/<intermediate_directories>/<artifact_filename>@
<build_number>",
where <intermediate_directories> can be one or a sequence of more
dirs where the artifacts are stored by a Jenkins job. Since the presence
and names of the directories can be arbitrary, it makes it difficult to
parse the purl to extract the build path only. Changed it to:
"pkg:<intermediate_directories>/<artifact_filename>@<build_number>?
build_path=<build_path>",
where the build path is separated and can be easily extracted.