Remove sending of EiffelArtifactCreatedEvent messages
It doesn't make sense for this plugin to send EiffelArtifactCreatedEvent
(ArtC) for each artifact. The package URLs produced by the plugin were
perhaps syntactically okay but semantically they weren't usable. Only
the build that produces the files knows what the ArtC events should look
like. This closes #8.

This is a backwards-incompatible change so let's bump the plugin version
from 1.0.2-SNAPSHOT to 2.0.0-SNAPSHOT.