Add basic tests for emitted events
Add a few tests for verifying that we get the events we expect for
a few scenarios and job types. Until we've refactored the event
classes to contain discrete members for all fields we don't actually
check the event contents but only that we get the expected number
of events.

Even that simple test uncovered a bug; we unintentionally sent an ActT
event for each step in a pipeline build. This bug will be addressed in
a follow-up commit.

To capture events sent via the MQConnection class we steal Mocks.java
from https://github.com/jenkinsci/mq-notifier-plugin at commit be56184
and make some adaptations.