Rewrite events to use Jackson-serialized POJOs; improve tests
The previous event classes that stored the event data in anonymous
hashmaps had a few problems:

    - IDEs weren't able to offer any help with the event construction.
    - Tests that wanted to inspect the resulting events would have to
      resort to lots of typecasting and try...catch cruft to to deal
      with the nested JSON objects.
    - Changing the event classes to make it easier to inspect the event
      data wouldn't help since there's no way to deserialize the JSON
      strings into objects of the original event classes.

This is solved by rewriting the event classes to plain POJOs and
relying on the Jackson library to transform instances to/from JSON.
This required a fair amount of code but the bulk of it is boilerplate
generated by the IDE.

This change allowed the tests to be rewritten and expanded to actually
pay attention to the event payloads.

The introduction of the Jackson dependency required an upgrade of
animal-sniffer-maven-plugin.