pom.xml: Upgrade core version and some deps and remove other deps
We upgrade the parent POM to support newer Jenkins releases and set
a few properties to match this. The Jenkins version used when testing
is explicitly set to 2.164.3 which is reasonably old and seems to be
common among widely-used plugins. Hopefully it's safe to assume that
things will work equally well on newer core versions. Previously we
got 1.6.something from the parent POM and that's really old.

Various dependencies are upgraded to the newest available version
to remain compatible with the newer core release (and avoid
RequireUpperBoundDeps errors) and a number of unused dependencies
are purged.