QueueListenerImpl: Don't emit ActT for pipeline steps
Apparently Jenkins calls QueueListener.onEnterWaiting for pipeline steps
too, so when starting a pipeline for the "foo" job we'll emit two ActT
events; one for "foo" and one for "part of foo #N". We fix this by
bailing out unless the Queue.WaitingItem.task implements BuildableItem.
There might be better ways of doing this but it seems to work fine.