Add trigger information to the ActT events
When a build is enqueued, populate the ActT event's data.trigger
attribute with objects translated from the build's causes. We currently
support TimerTriggerCause, SCMTriggerCause, and UpstreamCause. The first
two ones are quite straightforward and only involves setting the trigger
type (to TIMER and SOURCE_CHANGE, respectively) but the UpstreamCause
requires some work.

Jenkins can tell us which build triggered the current build, but what
we're interested in is the ID of that build's ActT event since that's
what should en up in the CAUSE link. We introduce a new action for this,
EiffelActivityTriggerAction, which also makes the ActT event available
to the outside world (via e.g. the $BUILD_URL/api/json endpoint).