Add EiffelCause and $JOB_URL/eiffel/build API endpoint
A new EiffelCause cause containing Eiffel event links is recognized
by the queue listener and populates the ActT event.

The new $JOB_URL/eiffel/build API endpoint complements the regular
$JOB_URL/build endpoint but also recognizes an "eiffellinks" key that,
if specified, gets parsed and ends up in the EiffelCause passed to
the started build. This allows clients to add e.g. CONTEXT and CAUSE
links to the Jenkins-originating activities.

There's one known bug/limitation: The EiffelCause class doesn't @Export
the list of links, only the short description. This shouldn't be a
serious problem since the equivalent information is available from the
complete events stored in the EiffelActivityAction that's also connected
to the run.