Populate meta.source.{host,name,uri}
We populated meta.source.name earlier too but with a made-up string
that wasn't used anywhere else. Use the plugin's long name instead
(currently "Eiffel Broadcaster Plugin").

To avoid Jenkins dependencies in the Eiffel POJOs as well as not
having to make separate calls to populate the meta.source members
each time an event instance is created we let the event package
define a SourceProvider interface that clients can set to an
environment-specific implementer. A SourceProvider reference is
stored in a static member in the EiffelEvent class. Perhaps not
super elegant but probably better than many other options.