Use the Run's EiffelActivityAction when linking ActF to ActT
Now that all Runs have an EiffelActivityAction we should use it to map
a completed Run to its ActT event instead of looking up the queue id
in EiffelJobTable. The latter isn't persisted so if the Jenkins master
restarts any pipeline build that survives the restart previously
wouldn't have gotten an ActF. Well, actually we would've attempted to
create an EiffelActivityFinishedEvent but the target of the ACTIVITY
link would've been null, probably resulting in an NPE somewhere.