Add job property for configuring activity categories
The new EiffelActivityJobProperty class can hold job-level settings that
affect the contents of the emitted Eiffel activities. To start with we
introduce a textarea that configures which categories are included in
the ActT. These categories are merged with the previously introduced
global categories.

Because the "properties" step in scripted pipelines is defined in the
workflow-multibranch plugin we had to add it as a test dependency. It
doesn't seem like the current version of that plugin is made available
via the parent POM so we specify a compatible version explicitly.

Bumping the minor plugin version in pom.xml to account for this new
feature.