Don't include URL schemes array in serialized configuration
The string array containing the names of the acceptable URL schemes,
EiffelBroadcasterConfig#schemes, was marked as "private final" which
caused it to be serialized to the plugin's configuration file. No big
deal, but it's wasteful and potentially confusing since the values
aren't really configurable.

We make that attribute transient (i.e. XStream can deserialize into
the attribute but will ignore it when serializing) and introduce a
new ALLOWED_URL_SCHEMES constant that won't be serialized (since it's
static).