Validate outbound events
For general sanity but also to act as a safety mechanism for the future
pipeline step for publishing an event we validate all outbound events
against the available schemas. Events that fail the validation won't be
published (but their payload will be logged).

For now only the schemas bundled with the plugin are available but we
can easily extend this to allow plugins to contribute schema provider
that can be centrally configurable. That way sites can configure this
plugin to fetch schemas via e.g. HTTP. The primary reason of that would
be to avoid having to update the plugin just to get access to a new
event or event version. We currently bundle all events and versions
available in https://github.com/eiffel-community/eiffel at the
edition-paris tag.

The plan was to use com.github.java-json-tools:json-schema-validator
but it needs Guava 12 and Jenkins only includes Guava 11. To avoid
dependency hell we went with com.networknt:json-schema-validator instead.
It also uses Jackson so the API is nearly identical, its performance is
allegedly better but the JavaDoc is quite poor. There were dependency
problems there as well so we went with a release from 2020 to avoid
depending on a very recent Jenkins core.