Allow arbitrary Eiffel events to be deserialized
To prepare for future changes where we need to deserialize arbitrary
Eiffel events and not just the ones we produce ourselves and have POJO
classes for we introduce a GenericEiffelEvent class that's used as the
fallback class if none of the existing POJO classes match. Previously
there was no fallback at all and an unknown event type would result in
an UnsupportedEventTypeException being thrown.

For GenericEiffelEvent we just save the data member in a JsonNode so
that the data is made available for introspection and kept for future
serialization back to JSON.

The addition of GenericEiffelEvent prompted a few other changes:

    - EiffelEvent.Meta.{id,time} are now mutable to support the custom
      initialization of GenericEiffelEvent.
    - We verify that the type name and version passed to the
      EiffelEvent.Meta constructor are set before allowing an object
      to be instantiated.
    - The EiffelEvent.Link.Type enumeration now covers all link types
      available as of the Paris edition of Eiffel. To support custom
      events we might want to make the link type a plain string instead
      of an enum.