Add sendEiffelEvent pipeline step
This new pipeline step allows sending arbitrary Eiffel events through
pipeline code. In addition to just enqueueing the supplied Groovy map it
by default adds a CONTEXT link to the Run's ActT event. The exact
behavior is configurable.

We extracted new Util.mustPublishEvent() methods from Util.publishEvent()
to allow for exceptions to be received by callers while the old
Util.publishEvent() is retained with the old behavior or just logging
any serialization or schema validation errors.

The new step shows up in the pipeline snippet generator but doesn't work
in a reasonable way. It turned out to be non-trivial to generate a
reasonable snippet given the mandatory Map-typed argument to the step.
Hopefully we can get that right in a follow-up commit.