Contruct and send ArtP events based on ArtC events
If ArtC and ArtP are sent together during the build, consumers are
susceptible to races since the Jenkins artifact URIs in the ArtP
might not be available (since artifact archival is usually done
at the very end of the build).

The new publishArtifact argument to the sendEiffelEvent step records
the ArtC in the Run and a subsequent call to publishEiffelArtifact when
the Jenkins artifacts have been made available (typically by using the
archiveArtifacts step) will construct and send ArtP event(s) based on
the ArtC events.