Allow publishEiffelArtifacts to read events from a file
Sometimes the artifact creations are buried deep inside build scripts
and calling the sendEiffelEvent pipeline step with publishArtifact set
isn't possible.

To support that use case we add an artifactEventFiles argument to the
publishEiffelArtifacts step. It contains an Ant-style glob expression
that selects workspace files from which to read JSON representations
of ArtC events. Those events will be treated exactly the same as events
recorded in the Run via sendEiffelEvent.