Support events from the Lyon edition of the protocol
We add the schemas introduced in the protocol's Lyon edition so that
users of the sendEiffelEvent pipeline step can send the new events
without getting stopped by the validator. The events emitted by the
plugin itself will stick to the Paris edition for now.

Apart from just adding the schema files for the validation we also
had to add class attributes for the event fields introduced in Lyon
for the events where we have corresponding classes. Otherwise added
event fields would get lost upon deserialization.