Populate meta.source for events sent via sendEiffelEvent step
Previously the meta.source fields were only populated when an event
class was instantiated directly, i.e. if the event was created by code
in the plugin. Event objects created by the sendEiffelEvent pipeline
step never asked the configured SourceProvider to populate meta.source.

The reason for this bug is that it was the EiffelEvent class that called
the SourceProvider, but this was only done in the constructor. When
deserializing JSON into an object the EiffelEvent#meta attribute is
overwritten by Jackson after the event object is initialized, hence
overwriting the SourceProvider-created value with the (possibly empty)
EiffelEvent.Meta object created by Jackson.

We addresss this by changing the custom deserializer to ask EiffelEvent
to populate meta.source after an event object has been created. This
also required changing JenkinsSourceProvider to not overwrite already
set meta.source values.