Make method of determining controller hostname configurable
Previously the hostname included in the meta.source.host member was
the hostname reported by the operating system, run through a name
resolution round trip. This wasn't necessarily the best choice if
the name the Jenkins controller was known under is a DNS alias. We
introduce a backwards compatible configuration option that chooses
between the old behavior (which is also the default) and grabbing
the hostname from the configured controller URL.