EiffelBroadcasterConfig: Extend GlobalConfiguration, add JCasC support
As a cleanup we convert EiffelBroadcasterConfig from extending
the deprecated Plugin class to GlobalConfiguration and make the
few adjustments required by that change, including:

    - Using a different way of obtaining the plugin's Plugin or
      PluginWrapper instances since Jenkins#getPlugin only works
      for plugins that extend Plugin.
    - Renaming the getter for EiffelBroadcasterConfig#enableBroadcaster
      so JCasC can find it.

Another complication of the switch from Plugin to GlobalConfiguration
was that the assignment to the static EiffelEvent#sourceProvider
attribute for some reason wasn't reset as it used to be. After a
test that used JenkinsRule was run the attribute still pointed to
a JenkinsSourceProvider instance that required a Jenkins instance to
be present. I'm not sure exactly what happens, but introducing
JenkinsRule subclasses that clear out EiffelEvent#sourceProvider after
each testcase fixes the problem.

The proper fix would be to not use static attributes like this, but
absent an injection framework like Guice it's not immediately obvious
how to do this (at least not without adding too many dependencies to
the rest of the plugin in EiffelEvent and its subclasses).