Upgrade Jenkins and parent POM
The minimum required Jenkins version is updated to 2.303.3 (which is
about a year old), and the BOM reference is updated accordingly. The
parent POM is updated to the latest possible given the Jenkins version.
Also, more plugins are now available in the BOM so a number of version
references could be removed.

Finally, the plugin description must now reside in index.jelly; not
the <description> element in the POM nor index.groovy. As per the
current recommendations the description in the POM is removed, but
that description text was much better than the one in index.groovy
so let's copy it into the newly created index.jelly.