Use @Terminator to clear out EiffelEvent's source provider
EiffelEvent unfortunately has a static source provider attribute
that populates the meta.source field in events. In JenkinsRule-based
integration tests we want this be work as it works in production,
adding a dependency to a running Jenkins instance. But after the
integration tests we might run unit tests, and they'll crash when
the Jenkins instance is no longer available.

Previously we'd subclassed JenkinsRule to clear out the source provider
when the test completes, but that's an unnecessary hack. Those two
subclasses are now replaced by a single class with a @Terminator method.