Fix startup race condition that put the outbound message queue in a bad state
For some time there's been a race condition during Jenkins startup
between the sending of the first event and the initialization of
the message queue thread.

The AMQP configuration is passed to MQConnection via initialized(),
called from EiffelBroadcasterConfig#configure(). Until this has
happened, the MQConnection's AMQP URL will be null and any attempt to
send an event will trigger an NPE. To make matters worse, the message
queue thread startup code in MQConnection#addMessageToQueue() had a
bug it would attempt to restart a crashed thread rather than create
a new one, resulting in an IllegalStateException.

We solve this by introducing an MQConnection#initialized attribute
that's set to true only after we've received the configuration, and
until then the message queue thread won't be started. We still post
messages to the queue, though, so any early messages will still be
sent.

While we're at it we solve the bug that resulted in the
IllegalStateException, allowing a crashed message queue thread to
be recoverable.