ConnectionIntegrationTest: Test concurrent publishing
To reduce the risk of deadlock bugs and lost message bugs slipping into
the code we add a testcase that uses a number of threads (currently 50)
to publish a sizeable chunk of events (1000) and asserts that they
arrive safely.

This adds a fair amount of test execution time and it's possible that
we're going to regret this later on. If that happens it might make sense
to keep the test but decrease the batch size. That way it's easy to run
the test locally on hopefully performant hardware (possibly adjusting
timeouts too) while still having stable CI including a more forgiving
test.