Support events from the Orizaba edition of the protocol
We add the schemas introduced in the protocol's Orizaba edition so
that users of the sendEiffelEvent pipeline step can send the new
events without getting stopped by the validator. The events emitted
by the plugin itself will continue to stick to the Paris edition.

Apart from just adding the schema files for the validation we also
had to add two newly introduced link types, CONFIGURATION and
ORIGINAL_TRIGGER, to the EiffelEvent.Link.Type enum.