Add support for signing system events
The optional signing is based on the signing method defined in the
Eiffel protocol. Eventually we want to be able to sign all kinds of
events, but this commit only adds support for signing system events,
i.e. events sent by the plugin itself and not under user control.

The private key used to sign an event is stored in a certificate
credential, where the certificate also contributes the subject,
which goes into the meta.security.authorIdentity field and helps
recipients choose the right public key for verifying the signature.

Credential tracking hasn't been implemented. This will be done in
a follow-up commit. We're going to need the job context anyway to
be able to resolve credentials properly when user events sent via
the sendEiffelEvent pipeline step are able to choose their own
credentials.