Create new JenkinsRule subclass to reduce duplication
We had duplicated code for creating a pipeline job and loading the
pipeline code from a resource file. This wasn't too bad, but in a
follow-up commit we're planning to create such jobs in subfolders,
which would make the problem worse. To reduce future pain we extract
the code to methods in a new JenkinsRule subclass.

Its name, JobCreatingJenkinsRule, isn't very imaginative but maybe
it's descriptive. If we want to widen the scope in the future it's
easy enough to rename it.