Add support for signing user events (#109)
In commit e9597aa we added support for signing system events, i.e.
events sent by the plugin itself and thus not under user control.
Here we add parameters to the sendEiffelEvent pipeline step to allow
job owners to upload their own credentials and configure their jobs
to use them for signing outbound events.

The design of the Util and SigningKeyCache classes had to change
when introducing signing of users events since a) obtaining system
credentials works differently than obtaining credentials from a folder,
and b) we can't use the ID of the credential as the cache key since
it would open up a security hole. We now always use the credentials
provider to resolve the credential for us, and the credential object
becomes the cache key. The cache still makes sense since it's the
extraction of the public key and subject DN that we want to avoid
doing too often.