Add buildWithEiffel pipeline step
This is a new pipeline step that extends BuildTriggerStep from
jenkinsci/pipeline-build-step-plugin#112 in order to add a new build
action EiffelActivityDataAction to the downstream build. This build 
action will supplement Eiffel activity data that will override
corresponding data fields in the EiffelActivityTriggeredEvent when
the downstream build enters the build queue.

Unfortunately, in order to add a new action to triggered builds,
the step execution of BuildTriggerStep had to be modified, which
meant that a number of project files had to be copied with slight
modifications to ensure that there is no interference when running
with both jenkinsci/eiffel-broadcaster-plugin and
jenkinsci/pipeline-build-step-plugin. Also, to ensure the underlying 
build trigger functionality inherited by the buildWithEiffel step,
a regression test suite has been added, containing all the
BuildTriggerStep tests. These tests have been updated to use
buildWithEiffel and the suite itself is excluded from normal testing
to reduce execution time.

Currently, buildWithEiffel step only supports activity name.