pom.xml: Upgrade parent POM and base Jenkins version to 2.426.3
Apart from being generally sane, it'll allow us to upgrade
com.networknt:json-schema-validator to a version that supports
draft 2020-12 of the JSON schema metaschema.

We were also able to remove a no longer necessary version override
for workflow-cps that was introduced in commit ac48793, plus an
orphaned property (jenkins-test-harness.version).