Introduce EiffelEventFactory
Right now each event class hardcodes its version. We want to support
central configuration of the Eiffel edition to use for outbound events,
and to avoid having a dependency to that configuration from each event
class we introduce a factory class. Right now it just has a hardcoded
event table, but that'll change in a follow-up commit.

We also move the SourceProvider from EiffelEvent to EiffelEventFactory,
making the event classes pure POJOs.

The drawback of the factory is that the classes are dumb in the sense
that their constructors no longer take parameters for mandatory fields,
i.e. it's less obvious what input you really need to provide to get a
valid event in the end.