Replace inheritance in BuildWithEiffelStep
Replaced inheritance of BuildTriggerStep and its descriptor with
modified copied code from the class itself to avoid breaking
the pipeline step document generator. Inheritance of a descriptor from
another step is not supported by the generator and we might as well
get rid off the Step inheritance to remove any dependencies
towards BuildTriggerStep.