Update copied code from pipeline-build-step-plugin from latest tag
Version 539.v8c889169451f added DownstreamBuildAction that also came
with a change to the Run listener when builds are started, which
would add/change this action on the build, and then save the build.

Some concurrency errors have been observed since then when running
buildWithEiffel step, although not in the tests.

To hopefully remedy this, the copied code has been updated to use
version  540.vb_e8849e1a_b_d8 of the build step plugin that includes
a couple of bug fixes for DownstreamBuildAction.