Adapt to Java 17 by removing mocking
To support Java 17, we had to do something about JMockIt's requirement
to modify the Surefire arguments. Instead of doing that we simply
removed the MQConnection mocking since it wasn't necessary. The tests
need to get hold of the events the plugin tried to send during the
test case, and MQConnection already tracks those messages.

To facilitate clearing the outbound queue in preparation of each test
case we introduce a JUnit rule.